requirements.txt
git+https://github.com/nocko/alchimia.git@pypy-fix
alembic==0.8.8
Jinja2==2.8
klein==15.3.1
psycopg2cffi==2.7.5
pytz==2016.7
Twisted==16.5.0
pycryptodome==3.4.3
enum-compat
